[FEAT] Extract battle event handlers
